\name{All}
\alias{All}
\docType{data}
\title{
    Available wrappers in the package
}
\description{
    Character vector containing all the names of the 
    available wrappers in the package
}
\usage{data("All")}
\format{
  Character vector containing the names.
}
\examples{
    data(All)
    print(All)
}
\keyword{datasets}
