\name{Fast}
\alias{Fast}
\docType{data}
\title{
    Available wrappers in the package of the fast methods
}
\description{
    Character vector containing  the names of the 
    wrappers in the package of the fast methods.
}
\usage{data("Fast")}
\format{
  Character vector containing the names of the fast methods.
}
\examples{
    data(Fast)
    print(Fast)
}
\keyword{datasets}
