\name{ntb_globals}
\alias{ntb_globals}
\docType{data}
\title{
    Available wrappers in the package of the fast methods
}

\description{
    Environment containing 
    a character vector containing  the names of the     
    wrappers in the package of the registered methods.
}

\usage{ntb_globals}

\format{
    Character vector containing the names of the registered
    wrapper methods.    
}

\examples{
    print(ntb_globals$Fast)
    print(ntb_globals$All)
}

\keyword{datasets}